backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
alembic
langchain
langchain-openai
langchain-google-genai
llama-index
qdrant-client
python-dotenv
httpx
pdfplumber
pytesseract
pillow
pytest
fpdf2
